Add BoxTextField
