+++ Projects/App/Sources/Present/Auth/SignUp/RegistrationPhoneNum/ViewController/RegistrationPhoneNumberViewController.swift
@@ -284,6 +284,6 @@ extension RegistrationPhoneNumberViewController: UITextFieldDelegate {
     }
     
     func textFieldDidEndEditing(_ textField: UITextField) {
-        textField.layer.borderColor = ChoiceAsset.Colors.grayMedium.color.cgColor
+        textField.layer.borderColor = UIColor.quaternaryLabel.cgColor
     }
 }
